requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.16.0
scrapy==2.11.0
playwright==1.40.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Geocoding & Maps
geopy==2.4.1
googlemaps==4.10.0

# API Clients
pyyaml==6.0.1
python-dotenv==1.0.0

# Rate Limiting & Caching
ratelimit==2.2.1
redis==5.0.1
cachetools==5.3.2

# Data Validation
pydantic==2.5.3
phonenumbers==8.13.27

# Dashboard
flask==3.0.0
jinja2==3.1.3

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
loguru==0.7.2

# Image Processing (for Street View)
pillow==10.2.0

# HTTP Client Alternatives
httpx==0.26.0
aiohttp==3.9.1

# Browser Automation
undetected-chromedriver==3.5.4

# Data Export
openpyxl==3.1.2
tabulate==0.9.0
